Phase 5 OS improvements: security hardening, shell reentrancy, IPC, VMM, FS
Security (5):
- Add argument validation for SYS_CHDIR/SYS_GETCWD/SYS_TRUNCATE/SYS_RAW_SEND
- Add argument validation for SYS_FD_READ/SYS_FD_WRITE
- Fix 2 pre-existing (uint64_t)-5 overflow warnings
- Block user-mode signal handler installation (privilege escalation)
- Signal_check now runs with interrupts disabled (bitmap integrity)

Shell reentrancy (7):
- Convert 7 static buffers in process_cmd to stack allocation (expanded,
  alias_expanded, glob_out, pipe_work, iredir_buf, iredir_args, redir_buf)
- Convert gnames glob buffer from static to auto
- Fix INT64_MIN arithmetic overflow (negating -2^63)
- Bound array index parsing in

Process/Scheduler (5):
- process_fork wrapped in cli/sti (slot allocation race)
- kmalloc moved before state=READY in fork (eliminates partial-init slot)
- waitpid re-looks up child after wake (fixes two-waiter race)
- signal_check removed from schedule -> moved to scheduler_tick (prevents
  recursive schedule() via signal delivery)
- process_get_by_pid prefers live over zombie (PID aliasing)

IPC/Pipes (3):
- Pipe multi-waiter support (arrays of 4 PIDs instead of single PID)
- Pipe negative len returns -1
- mutex_init cli/sti for atomic slot allocation

VMM (3):
- vmm_virt_to_phys with proper error signaling (0=ok, -1=not mapped)
- vmm_switch_pml4/destroy use VIRT_TO_PHYS (not identity assumption)
- vmm_user_string_ok TOCTOU fixed (cli/sti around check+deref)

FS (4):
- fs_read_file skips unallocated blocks (zero-fills instead of reading sector 0)
- fs_truncate allocates blocks when extending file
- fs_write_file crash ordering: update inode before freeing old blocks
- find_inode symlink depth tracked globally (not reset on recursive call)
- find_inode('/') returns inode 0 (root directory)

FAT32 (4):
- mkdir creates '.' and '..' directory entries
- Cluster leak fixed on dir_add_entry failure
- FAT chain cycle detection (max clusters bound)
- DNS cache dedup: checks existing name before allocating new slot

Drivers (4):
- ATA uses __builtin_memcpy instead of misaligned uint16_t* cast
- ATA identify timeout: bsy wait return checked, pause in DRQ loop
- e1000 RX: checks errors and EOP bits (drops corrupted/fragmented)
- Keyboard ring buffer dequeue in cli/sti

TCP (5):
- FIN_WAIT->FIN_WAIT_2 on ACK; FIN_WAIT_2->TIME_WAIT on FIN
- CLOSE_WAIT->LAST_ACK->CLOSED (proper close handshake)
- Accept queue overflow sends RST
- Retransmit chunks into MSS-sized segments (no IP frag)
- TCP zero-checksum rejected (RFC 793 compliance)

Heap (1):
- Heap coalescing no longer underflows heap_used_bytes

Tests (4):
- test_heap_ext: coalescing accounting verification
- test_pipe: negative len edge cases
- test_signal: mask/unmask test
- VFS path resolution tests (root '/', '..' resolution)

25 files, +615 -208, zero warnings on both normal and test builds

+++ src/net/net.c
@@ -394,7 +394,7 @@ static void handle_ip(const uint8_t *data, uint16_t len);
 
 static int handle_ip_fragment(struct ip_header *ip, const uint8_t *data, uint16_t len) {
     uint16_t flags_frag = ntohs(ip->flags_frag);
-    uint16_t frag_off = (flags_frag & 0x1FFF) * 8;
+    uint32_t frag_off = (uint32_t)(flags_frag & 0x1FFF) * 8;
     int more = (flags_frag & 0x2000) != 0;
 
     if (frag_off == 0 && !more)
@@ -407,7 +407,7 @@ static int handle_ip_fragment(struct ip_header *ip, const uint8_t *data, uint16_
 
     uint16_t ihl = (ip->version_ihl & 0xF) * 4;
     uint16_t part = len - ihl;
-    if (frag_off + part > sizeof(slot->buf)) return -1;
+    if (frag_off + (uint32_t)part > sizeof(slot->buf)) return -1;
 
     /* Reject overlapping fragments (security: prevent data injection attacks) */
     for (uint16_t b = frag_off; b < frag_off + part; b++) {
@@ -416,7 +416,7 @@ static int handle_ip_fragment(struct ip_header *ip, const uint8_t *data, uint16_
     }
 
     memcpy(slot->buf + frag_off, data + ihl, part);
-    for (uint16_t b = frag_off; b < frag_off + part; b++)
+    for (uint32_t b = frag_off; b < frag_off + part; b++)
         slot->frag_map[b / 8] |= (uint8_t)(1u << (b % 8));
     if (frag_off + part > slot->len) slot->len = (uint16_t)(frag_off + part);
     if (frag_off + part > slot->frag_end) slot->frag_end = (uint16_t)(frag_off + part);
